Admin HTML
Added the html/css for the admin functions. Right now whether or not it
shows is just a hardcoded variable at the top of nav.php. Once we get
the logins working we can then check to see if a user is 1) logged in 2)
if they are an admin. If they are an admin then set $admin = true and
the admin controls will be displayed.

I have not implemented any of the admin controls, but I populated the
forms for adding a new book, new author, or new subgenre.

I still need to create the function to populate the form with the
correct information for editing a book. and of course I need to create
the scripts for submitting the forms.

I think I'm going to see if I can change the ids for subgenres and
authors to autoincrement, but I don't know if it will work with the
database already being filled.