Form validation for adding book
Working on the form validation for adding a book.

Still need to actually insert the data into the tables and figure out
how to upload an image file.

Currently the form uses jquery to validate and uses ajax to post all
data to a php file. It currently just prints the information to the
console.