Edit Book Function
The edit book button now opens up the updatebook form and fills the data
in. I have not implemented actually updating the book in the database.
And I would like to add labels of some sort to the form so that you can
tell what each value is.

I also moved the book printing to a seperate php file because it was on
two different pages and kept getting messed up between the index and
filter page.